Experiment with persisting state across page reloads
