+++ tools/src/lib/persistedStore.ts
@@ -1,0 +1,26 @@
+import { writable } from 'svelte/store'
+
+export default function persistedStore<T>(key: string, startValue: T) {
+    const { subscribe, set } = writable(startValue)
+
+    return {
+        subscribe,
+        set,
+        useLocalStorage: () => {
+            const json = localStorage.getItem(key)
+            if (typeof json === 'string') {
+                let data: T
+                try {
+                    data = JSON.parse(json)
+                } catch (error) {
+                    data = startValue
+                }
+                set(data)
+            }
+
+            subscribe((current) => {
+                localStorage.setItem(key, JSON.stringify(current))
+            })
+        },
+    }
+}
